python-GetNotionData/requirements.txt
python-dotenv
notion-client
requests
mysql-connector-python